conv3d dequan X寫進去裡面

+++ R2+1D/r2plus1d.h
@@ -4,7 +4,7 @@
 #define N_ 1
 #define C_ 1152
 #define D_ 16
-typedef int dtype;
+typedef uint8_t dtype;
 typedef int int_t;
 
 void r2plus1d(dtype* X, dtype* Y, dtype* Kernel_stem_0, dtype* Kernel_stem_3, double* Kernel_stem_0_scale, double* Kernel_stem_3_scale,
